chore: reduce complexity
